Trait notion_fail::NotionFail

pub trait NotionFail: Fail {
    fn is_user_friendly(&self) -> bool;

    fn exit_code(&self) -> ExitCode;
}

The failure trait for all Notion errors.

Required methods

fn is_user_friendly(&self) -> bool

Indicates whether this error has a message suitable for reporting to an end-user.

fn exit_code(&self) -> ExitCode

Returns the process exit code that should be returned if the process exits with this error.

Loading content...